task#427: unify product-api
