+++ src/aws-tests/src/staging-tests/Cargo.toml
@@ -29,8 +29,7 @@ reqwest = { workspace = true, features = ["default", "json"] }
 common = { workspace = true, features = ["test-data"] }
 serial_test = { workspace = true }
 http = { workspace = true }
-product-api-watchlist-get = { workspace = true }
-product-api-watchlist-patch = { workspace = true }
+product-api = { workspace = true }
 search-filter-api-patch-search-filter = { workspace = true, features = ["test-data"] }
 search-filter-api-post-search-filter = { workspace = true, features = ["test-data"] }
 url = { workspace = true }
