--- src/product-api/src/product-api-get-product-similar/Cargo.toml
@@ -1,29 +1,0 @@
-[package]
-name = "product-api-get-product-similar"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-common = { workspace = true, features = ["api", "opensearch"] }
-product = { workspace = true, features = ["data", "service", "opensearch", "watchlist", "dynamodb"] }
-cognito = { workspace = true }
-tracing = { workspace = true }
-aws-sdk-dynamodb = { workspace = true }
-tracing-subscriber = { workspace = true, features = ["json"] }
-tokio = { workspace = true, features = ["macros"] }
-aws_lambda_events = { workspace = true, features = ["apigw"] }
-lambda_runtime = { workspace = true }
-aws-config = { workspace = true }
-opensearch = { workspace = true }
-serde_json = { workspace = true }
-http = { workspace = true }
-
-[dev-dependencies]
-product = { workspace = true, features = ["test-data"] }
-user = { workspace = true, features = ["full"] }
-test-api = { workspace = true, features = ["api-gateway", "opensearch", "dynamodb"] }
-rstest = { workspace = true }
-serial_test = { workspace = true }
-http = { workspace = true }
-fake = { workspace = true }
-time = { workspace = true }
