--- src/product-api/src/product-api-get-product/Cargo.toml
@@ -1,29 +1,0 @@
-[package]
-name = "product-api-get-product"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-common = { workspace = true, features = ["api"] }
-product = { workspace = true, features = ["data", "service", "dynamodb", "watchlist"] }
-cognito = { workspace = true }
-tracing = { workspace = true }
-tracing-subscriber = { workspace = true, features = ["json"] }
-tokio = { workspace = true, features = ["macros"] }
-aws_lambda_events = { workspace = true, features = ["apigw"] }
-lambda_runtime = { workspace = true }
-aws-config = { workspace = true }
-aws-sdk-dynamodb = { workspace = true }
-serde_json = { workspace = true }
-http = { workspace = true }
-
-[dev-dependencies]
-product = { workspace = true, features = ["test-data"] }
-user = { workspace = true, features = ["test-data", "dynamodb"] }
-test-api = { workspace = true, features = ["api-gateway", "dynamodb"] }
-serial_test = { workspace = true }
-rstest = { workspace = true }
-time = { workspace = true }
-http = { workspace = true }
-url = { workspace = true }
-fake = { workspace = true }
