+++ src/product-api/src/watchlist_delete.rs
@@ -1,39 +1,12 @@
 use aws_lambda_events::apigw::{ApiGatewayV2httpRequest, ApiGatewayV2httpResponse};
 use common::api::api_gateway_v2_http_response_builder::ApiGatewayV2HttpResponseBuilder;
-use common::api::error::{ApiError, log_api_error};
+use common::api::error::ApiError;
 use common::shop_id::api::extract_shop_id_path;
 use common::shops_product_id::api::extract_shops_product_id_path;
 use common::user_id::api::extract_user_id_request_context;
 use lambda_runtime::LambdaEvent;
 use product::watchlist::service::product_watchlist_service::ProductWatchListService;
 
-#[tracing::instrument(
-    skip(event, service),
-    fields(
-        requestId = %event.context.request_id,
-        method = event.payload.request_context.http.method.as_str(),
-        path = &event.payload.raw_path.as_deref().unwrap_or("NULL"),
-        query = &event.payload.raw_query_string.as_deref().unwrap_or("NULL"),
-        body = &event.payload.body.as_deref().unwrap_or("NULL"),
-        ip = &event.payload.request_context.http.source_ip.as_deref().unwrap_or("NULL"),
-        userAgent = &event.payload.request_context.http.user_agent.as_deref().unwrap_or("NULL"),
-        userId = tracing::field::Empty,
-    )
-)]
-pub async fn handler(
-    event: LambdaEvent<ApiGatewayV2httpRequest>,
-    service: &impl ProductWatchListService,
-) -> Result<ApiGatewayV2httpResponse, lambda_runtime::Error> {
-    match handle(event, service).await {
-        Ok(response) => Ok(response),
-        Err(err) => {
-            log_api_error(&err);
-            Ok(ApiGatewayV2httpResponse::from(err))
-        }
-    }
-}
-
-// DELETE /api/v1/me/watchlist/{shopId}/{shopsProductId}
 pub async fn handle(
     event: LambdaEvent<ApiGatewayV2httpRequest>,
     service: &impl ProductWatchListService,
@@ -52,7 +25,7 @@ pub async fn handle(
 
 #[cfg(test)]
 mod tests {
-    use crate::handler;
+    use super::handle;
     use common::{shop_id::ShopId, shops_product_id::ShopsProductId, user_id::UserId};
     use lambda_runtime::LambdaEvent;
     use product::watchlist::service::product_watchlist_service::MockProductWatchListService;
@@ -80,7 +53,7 @@ mod tests {
             context: Default::default(),
         };
 
-        let response = handler(lambda_event, &service).await.unwrap();
+        let response = handle(lambda_event, &service).await.unwrap();
 
         assert_eq!(204, response.status_code);
     }
@@ -103,7 +76,7 @@ mod tests {
             context: Default::default(),
         };
 
-        let response = handler(lambda_event, &service).await.unwrap();
+        let response = handle(lambda_event, &service).await.unwrap();
         assert_eq!(400, response.status_code);
 
         let json = extract_apigw_response_json_body!(response);
@@ -130,7 +103,7 @@ mod tests {
             context: Default::default(),
         };
 
-        let response = handler(lambda_event, &service).await.unwrap();
+        let response = handle(lambda_event, &service).await.unwrap();
         assert_eq!(400, response.status_code);
 
         let json = extract_apigw_response_json_body!(response);
@@ -157,7 +130,7 @@ mod tests {
             context: Default::default(),
         };
 
-        let response = handler(lambda_event, &service).await.unwrap();
+        let response = handle(lambda_event, &service).await.unwrap();
 
         assert_eq!(401, response.status_code);
     }
