+++ src/product-api/src/watchlist_get.rs
@@ -1,5 +1,5 @@
 use aws_lambda_events::apigw::{ApiGatewayV2httpRequest, ApiGatewayV2httpResponse};
-use common::api::error::{ApiError, log_api_error};
+use common::api::error::ApiError;
 use common::currency::data::api::extract_currency_query;
 use common::language::data::api::extract_language_header;
 use common::pagination::cursor::api::{TimeCursoredData, extract_time_cursor_query};
@@ -42,33 +42,6 @@ impl From<LocalizedWatchlistProductView> for WatchlistProductDataView {
     }
 }
 
-#[tracing::instrument(
-    skip(event, service),
-    fields(
-        requestId = %event.context.request_id,
-        method = event.payload.request_context.http.method.as_str(),
-        path = &event.payload.raw_path.as_deref().unwrap_or("NULL"),
-        query = &event.payload.raw_query_string.as_deref().unwrap_or("NULL"),
-        body = &event.payload.body.as_deref().unwrap_or("NULL"),
-        ip = &event.payload.request_context.http.source_ip.as_deref().unwrap_or("NULL"),
-        userAgent = &event.payload.request_context.http.user_agent.as_deref().unwrap_or("NULL"),
-        userId = tracing::field::Empty,
-    )
-)]
-pub async fn handler(
-    event: LambdaEvent<ApiGatewayV2httpRequest>,
-    service: &impl ProductWatchListService,
-) -> Result<ApiGatewayV2httpResponse, lambda_runtime::Error> {
-    match handle(event, service).await {
-        Ok(response) => Ok(response),
-        Err(err) => {
-            log_api_error(&err);
-            Ok(ApiGatewayV2httpResponse::from(err))
-        }
-    }
-}
-
-// GET /api/v1/me/watchlist
 pub async fn handle(
     event: LambdaEvent<ApiGatewayV2httpRequest>,
     service: &impl ProductWatchListService,
@@ -101,7 +74,7 @@ pub async fn handle(
 
 #[cfg(test)]
 mod tests {
-    use crate::handler;
+    use super::handle;
     use common::user_id::UserId;
     use fake::{Fake, Faker};
     use lambda_runtime::LambdaEvent;
@@ -130,7 +103,7 @@ mod tests {
             context: Default::default(),
         };
 
-        let response = handler(lambda_event, &service).await.unwrap();
+        let response = handle(lambda_event, &service).await.unwrap();
 
         assert_eq!(200, response.status_code);
     }
@@ -151,7 +124,7 @@ mod tests {
             context: Default::default(),
         };
 
-        let response = handler(lambda_event, &service).await.unwrap();
+        let response = handle(lambda_event, &service).await.unwrap();
 
         assert_eq!(401, response.status_code);
     }
