+++ src/product-api/src/watchlist_patch.rs
@@ -2,7 +2,7 @@ use aws_lambda_events::apigw::{ApiGatewayV2httpRequest, ApiGatewayV2httpResponse
 use common::api::error::ApiError;
 use common::api::error_code::BAD_BODY_VALUE;
 use common::api::{
-    api_gateway_v2_http_response_builder::ApiGatewayV2HttpResponseBuilder, error::log_api_error,
+    api_gateway_v2_http_response_builder::ApiGatewayV2HttpResponseBuilder,
 };
 use common::shop_id::api::extract_shop_id_path;
 use common::shops_product_id::api::extract_shops_product_id_path;
@@ -31,33 +31,6 @@ impl From<WatchlistProductPatch> for UpdateWatchlistProductCommand {
     }
 }
 
-#[tracing::instrument(
-    skip(event, service),
-    fields(
-        requestId = %event.context.request_id,
-        method = event.payload.request_context.http.method.as_str(),
-        path = &event.payload.raw_path.as_deref().unwrap_or("NULL"),
-        query = &event.payload.raw_query_string.as_deref().unwrap_or("NULL"),
-        body = &event.payload.body.as_deref().unwrap_or("NULL"),
-        ip = &event.payload.request_context.http.source_ip.as_deref().unwrap_or("NULL"),
-        userAgent = &event.payload.request_context.http.user_agent.as_deref().unwrap_or("NULL"),
-        userId = tracing::field::Empty,
-    )
-)]
-pub async fn handler(
-    event: LambdaEvent<ApiGatewayV2httpRequest>,
-    service: &impl ProductWatchListService,
-) -> Result<ApiGatewayV2httpResponse, lambda_runtime::Error> {
-    match handle(event, service).await {
-        Ok(response) => Ok(response),
-        Err(err) => {
-            log_api_error(&err);
-            Ok(ApiGatewayV2httpResponse::from(err))
-        }
-    }
-}
-
-// PATCH /api/v1/me/watchlist/{shopId}/{shopsProductId}
 pub async fn handle(
     event: LambdaEvent<ApiGatewayV2httpRequest>,
     service: &impl ProductWatchListService,
@@ -91,7 +64,8 @@ pub async fn handle(
 
 #[cfg(test)]
 mod tests {
-    use crate::{WatchlistProductPatch, handler};
+    use super::WatchlistProductPatch;
+    use super::handle;
     use common::{shop_id::ShopId, shops_product_id::ShopsProductId, user_id::UserId};
     use fake::{Fake, Faker};
     use lambda_runtime::LambdaEvent;
@@ -123,7 +97,7 @@ mod tests {
             context: Default::default(),
         };
 
-        let response = handler(lambda_event, &service).await.unwrap();
+        let response = handle(lambda_event, &service).await.unwrap();
 
         assert_eq!(200, response.status_code);
     }
@@ -149,7 +123,7 @@ mod tests {
             context: Default::default(),
         };
 
-        let response = handler(lambda_event, &service).await.unwrap();
+        let response = handle(lambda_event, &service).await.unwrap();
         assert_eq!(400, response.status_code);
 
         let json = extract_apigw_response_json_body!(response);
@@ -179,7 +153,7 @@ mod tests {
             context: Default::default(),
         };
 
-        let response = handler(lambda_event, &service).await.unwrap();
+        let response = handle(lambda_event, &service).await.unwrap();
         assert_eq!(400, response.status_code);
 
         let json = extract_apigw_response_json_body!(response);
@@ -207,7 +181,7 @@ mod tests {
             context: Default::default(),
         };
 
-        let response = handler(lambda_event, &service).await.unwrap();
+        let response = handle(lambda_event, &service).await.unwrap();
         assert_eq!(400, response.status_code);
 
         let json = extract_apigw_response_json_body!(response);
@@ -234,7 +208,7 @@ mod tests {
             context: Default::default(),
         };
 
-        let response = handler(lambda_event, &service).await.unwrap();
+        let response = handle(lambda_event, &service).await.unwrap();
         assert_eq!(400, response.status_code);
 
         let json = extract_apigw_response_json_body!(response);
@@ -262,7 +236,7 @@ mod tests {
             context: Default::default(),
         };
 
-        let response = handler(lambda_event, &service).await.unwrap();
+        let response = handle(lambda_event, &service).await.unwrap();
 
         assert_eq!(401, response.status_code);
     }
