+++ src/product-api/src/watchlist_post.rs
@@ -1,6 +1,6 @@
 use aws_lambda_events::apigw::{ApiGatewayV2httpRequest, ApiGatewayV2httpResponse};
 use common::api::api_gateway_v2_http_response_builder::ApiGatewayV2HttpResponseBuilder;
-use common::api::error::{ApiError, log_api_error};
+use common::api::error::ApiError;
 use common::api::error_code::BAD_BODY_VALUE;
 use common::product_id::api::ProductKeyData;
 use common::user_id::api::extract_user_id_request_context;
@@ -8,33 +8,6 @@ use lambda_runtime::LambdaEvent;
 use product::watchlist::data::watchlist_product_data::WatchlistProductData;
 use product::watchlist::service::product_watchlist_service::ProductWatchListService;
 
-#[tracing::instrument(
-    skip(event, service),
-    fields(
-        requestId = %event.context.request_id,
-        method = event.payload.request_context.http.method.as_str(),
-        path = &event.payload.raw_path.as_deref().unwrap_or("NULL"),
-        query = &event.payload.raw_query_string.as_deref().unwrap_or("NULL"),
-        body = &event.payload.body.as_deref().unwrap_or("NULL"),
-        ip = &event.payload.request_context.http.source_ip.as_deref().unwrap_or("NULL"),
-        userAgent = &event.payload.request_context.http.user_agent.as_deref().unwrap_or("NULL"),
-        userId = tracing::field::Empty,
-    )
-)]
-pub async fn handler(
-    event: LambdaEvent<ApiGatewayV2httpRequest>,
-    service: &impl ProductWatchListService,
-) -> Result<ApiGatewayV2httpResponse, lambda_runtime::Error> {
-    match handle(event, service).await {
-        Ok(response) => Ok(response),
-        Err(err) => {
-            log_api_error(&err);
-            Ok(ApiGatewayV2httpResponse::from(err))
-        }
-    }
-}
-
-// POST /api/v1/me/watchlist
 pub async fn handle(
     event: LambdaEvent<ApiGatewayV2httpRequest>,
     service: &impl ProductWatchListService,
@@ -81,7 +54,7 @@ pub async fn handle(
 
 #[cfg(test)]
 mod tests {
-    use crate::handler;
+    use super::handle;
     use common::{product_id::api::ProductKeyData, user_id::UserId};
     use fake::{Fake, Faker};
     use http::header::LOCATION;
@@ -108,7 +81,7 @@ mod tests {
             context: Default::default(),
         };
 
-        let response = handler(lambda_event, &service).await.unwrap();
+        let response = handle(lambda_event, &service).await.unwrap();
 
         assert_eq!(201, response.status_code);
         assert_eq!(
@@ -133,7 +106,7 @@ mod tests {
             context: Default::default(),
         };
 
-        let response = handler(lambda_event, &service).await.unwrap();
+        let response = handle(lambda_event, &service).await.unwrap();
 
         assert_eq!(401, response.status_code);
     }
